requirements.txt
rembg==2.0.57
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.1.0
gunicorn==21.2.0
